rename `@tokenami/dev` to `tokenami`
