+++ .npmrc
@@ -4,4 +4,5 @@ auto-install-peers=true
 link-workspace-packages=true
 prefer-workspace-packages=true
 hoist-workspace-packages=true
-public-hoist-pattern[]=@tokenami/dev # needed for the ts-plugin to load locally
+public-hoist-pattern[]=tokenami # needed for the ts-plugin to load locally
+public-hoist-pattern[]=!@tokenami/*
