+++ examples/design-system/package.json
@@ -27,8 +27,8 @@
   "devDependencies": {
     "@storylite/storylite": "^0.14.0",
     "@storylite/vite-plugin": "^0.14.0",
-    "@tokenami/dev": "workspace:*",
     "@vitejs/plugin-react-swc": "^3.6.0",
+    "tokenami": "workspace:*",
     "tsup": "^7.0.0",
     "typescript": "^5.1.3",
     "vite": "^5.1.6"
