+++ examples/design-system/tsconfig.json
@@ -4,7 +4,7 @@
   "compilerOptions": {
     "baseUrl": ".",
     "tsBuildInfoFile": "node_modules/.cache/tsbuildinfo.json",
-    "plugins": [{ "name": "@tokenami/dev" }],
+    "plugins": [{ "name": "tokenami" }],
     "jsx": "react-jsx",
     "allowSyntheticDefaultImports": true,
     "types": ["node", "react", "react-dom", "@storylite/vite-plugin/virtual"],
