+++ examples/remix/.tokenami/tokenami.env.ci.d.ts
@@ -1,7 +1,7 @@
 /// <reference types="./tokenami.env.d.ts" />
-import type { Config } from './tokenami.env';
+import type { Config } from './tokenami.env.js';
 
-declare module '@tokenami/dev' {
+declare module 'tokenami' {
   interface TokenamiConfig extends Config {
     CI: true;
   }
