+++ examples/remix/package.json
@@ -26,10 +26,10 @@
   "devDependencies": {
     "@remix-run/dev": "^1.17.1",
     "@remix-run/eslint-config": "^1.17.1",
-    "@tokenami/dev": "workspace:*",
     "@types/react": "^18.0.35",
     "@types/react-dom": "^18.0.11",
     "eslint": "^8.57.0",
+    "tokenami": "workspace:*",
     "typescript": "^5.1.3"
   },
   "eslintIgnore": [
