+++ examples/remix/tsconfig.json
@@ -4,7 +4,7 @@
   "compilerOptions": {
     "baseUrl": ".",
     "tsBuildInfoFile": "node_modules/.cache/tsbuildinfo.json",
-    "plugins": [{ "name": "@tokenami/dev" }],
+    "plugins": [{ "name": "tokenami" }],
     "jsx": "react-jsx",
     "lib": ["DOM", "DOM.Iterable", "ES2019"],
     "esModuleInterop": true,
