+++ packages/css/.tokenami/tokenami.env.d.ts
@@ -2,7 +2,7 @@ import config from './tokenami.config';
 
 export type Config = typeof config;
 
-declare module '@tokenami/dev' {
+declare module 'tokenami' {
   interface TokenamiConfig extends Config {}
   interface TokenamiProperties {}
 }
