+++ packages/css/src/index.ts
@@ -1,4 +1,4 @@
-import type { TokenamiProperties } from '@tokenami/dev';
+import type { TokenamiProperties } from 'tokenami';
 import type { TokenamiCSS } from './css';
 
 export type TokenamiStyle<P> = Omit<P, 'style'> & {
@@ -9,6 +9,6 @@ export type Variants<T extends (...args: any) => any> = Parameters<T>[0] extends
   ? {}
   : NonNullable<Parameters<T>[0]>;
 
-export type { TokenamiProperties } from '@tokenami/dev';
+export type { TokenamiProperties } from 'tokenami';
 export { type Config, createConfig } from '@tokenami/config';
 export { type TokenamiCSS, type CSS, createCss, css } from './css';
