+++ packages/css/tsconfig.json
@@ -8,7 +8,7 @@
     "target": "ES2022",
     "paths": {
       "@tokenami/config": ["../config/src"],
-      "@tokenami/dev": ["../dev/src"]
+      "tokenami": ["../tokenami/src"]
     }
   }
 }
