+++ packages/tokenami/README.md
@@ -1,3 +1,3 @@
-# @tokenami/dev
+# tokenami
 
 View docs [here](https://github.com/tokenami/tokenami).
