+++ packages/tokenami/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@tokenami/dev",
+  "name": "tokenami",
   "version": "0.0.76",
   "license": "MIT",
   "type": "module",
