+++ packages/tokenami/src/declarations.ts
