+++ packages/tokenami/stubs/tokenami.env-custom.d.ts
@@ -1,9 +1,9 @@
-import { type TokenProperties } from '@tokenami/dev';
+import { type TokenProperties } from 'tokenami';
 import config from './tokenami.config.js';
 
 export type Config = typeof config;
 
-declare module '@tokenami/dev' {
+declare module 'tokenami' {
   interface TokenamiConfig extends Config {}
   interface TokenamiProperties {}
 }
