+++ packages/tokenami/stubs/tokenami.env.d.ts
@@ -2,6 +2,6 @@ import config from './tokenami.config.js';
 
 export type Config = typeof config;
 
-declare module '@tokenami/dev' {
+declare module 'tokenami' {
   interface TokenamiConfig extends Config {}
 }
